+++ developer/src/kmc-ldml/src/compiler/disp.ts
@@ -27,8 +27,8 @@ export class DispCompiler extends SectionCompiler {
     const tos = new Set();
     const ids = new Set();
 
-    if (this.keyboard.displays?.display) {
-      for (const { to, id } of this.keyboard.displays?.display) {
+    if (this.keyboard3.displays?.display) {
+      for (const { to, id } of this.keyboard3.displays?.display) {
         if ((to && id) || (!to && !id)) {
           this.callbacks.reportMessage(CompilerMessages.Error_DisplayNeedsToOrId({ to, id }));
           return false;
@@ -57,15 +57,20 @@ export class DispCompiler extends SectionCompiler {
     let result = new Disp();
 
     // displayOptions
-    result.baseCharacter = sections.strs.allocAndUnescapeString(this.keyboard.displays?.displayOptions?.baseCharacter);
-
-    // TODO-LDML: substitute variables!
+    result.baseCharacter = sections.strs.allocString(this.keyboard3.displays?.displayOptions?.baseCharacter, {unescape: true});
 
     // displays
-    result.disps = this.keyboard.displays?.display.map(display => ({
-      to: sections.strs.allocAndUnescapeString(sections.vars.substituteMarkerString(display.to)),
+    result.disps = this.keyboard3.displays?.display.map(display => ({
+      to: sections.strs.allocString(display.to, {
+        stringVariables: true,
+        markers: true,
+        unescape: true,
+      }, sections),
       id: sections.strs.allocString(display.id), // not escaped, not substituted
-      display: sections.strs.allocAndUnescapeString(display.display),
+      display: sections.strs.allocString(display.display, {
+        stringVariables: true,
+        unescape: true,
+      }, sections),
     })) || []; // TODO-LDML: need coverage for the []
 
     result.disps.sort((a: DispItem, b: DispItem) => {