+++ developer/src/kmc-ldml/src/compiler/keys.ts
@@ -30,15 +30,24 @@ export class KeysCompiler extends SectionCompiler {
    * @returns just the non-touch layers.
    */
   public hardwareLayers() {
-    return this.keyboard.layers?.filter(({ form }) => form !== "touch");
+    return this.keyboard3.layers?.filter(({ form }) => form !== "touch");
   }
 
   public validate() {
     let valid = true;
 
+    // There's no 'form' compiler.
+    // We validate this here so that someone checks it.
+    this.keyboard3.forms?.form?.forEach((form) => {
+      if (!LDMLKeyboard.ImportStatus.isImpliedImport(form)) {
+        // If it's not an implied import, give a warning.
+        this.callbacks.reportMessage(CompilerMessages.Warn_CustomForm({ id: form.id }));
+      }
+    });
+
     // general key-level validation here, only of used keys
-    const usedKeys = allUsedKeyIdsInLayers(this.keyboard?.layers);
-    const uniqueKeys = calculateUniqueKeys([...this.keyboard.keys?.key]);
+    const usedKeys = allUsedKeyIdsInLayers(this.keyboard3?.layers);
+    const uniqueKeys = calculateUniqueKeys([...this.keyboard3.keys?.key]);
     for (let key of uniqueKeys) {
       const { id, flicks } = key;
       if (!usedKeys.has(id)) {
@@ -48,7 +57,7 @@ export class KeysCompiler extends SectionCompiler {
       if (!flicks) {
         continue; // no flicks
       }
-      const flickEntry = this.keyboard.keys?.flicks?.find(
+      const flickEntry = this.keyboard3.keys?.flicks?.find(
         (x) => x.id === flicks
       );
       if (!flickEntry) {
@@ -80,7 +89,7 @@ export class KeysCompiler extends SectionCompiler {
 
   public compile(sections: DependencySections): Keys {
     /* c8 ignore next 4 */
-    if (!this.keyboard?.keys?.key && !this.keyboard?.keys?.flicks) {
+    if (!this.keyboard3?.keys?.key && !this.keyboard3?.keys?.flicks) {
       // short-circuit if no keys or flicks. Doesn't happen in practice due to implied import.
       return null;
     }
@@ -114,25 +123,28 @@ export class KeysCompiler extends SectionCompiler {
   }
 
   public loadFlicks(sections: DependencySections, sect: Keys) {
-    for (let lkflicks of this.keyboard.keys.flicks) {
+    for (let lkflicks of this.keyboard3.keys.flicks) {
       let flicks: KeysFlicks = new KeysFlicks(
         sections.strs.allocString(lkflicks.id)
       );
 
       for (let lkflick of lkflicks.flick) {
         let flags = 0;
-        const to = sections.strs.allocAndUnescapeString(lkflick.to, true);
+        const to = sections.strs.allocString(lkflick.to, {
+          stringVariables: true, markers: true, unescape: true, singleOk: true
+        }, sections);
         if (!to.isOneChar) {
           flags |= constants.keys_flick_flags_extend;
         }
         let directions: ListItem = sections.list.allocListFromSpaces(
-          sections.strs,
-          lkflick.directions
-        );
+          lkflick.directions,
+          {
+            stringVariables: true, markers: true, unescape: true
+          },
+          sections);
         flicks.flicks.push({
           directions,
           flags,
-          // TODO-LDML: markers,variables
           to,
         });
       }
@@ -142,8 +154,8 @@ export class KeysCompiler extends SectionCompiler {
   }
 
   public loadKeys(sections: DependencySections, sect: Keys) {
-    const usedKeys = allUsedKeyIdsInLayers(this.keyboard?.layers);
-    const uniqueKeys = calculateUniqueKeys([...this.keyboard.keys?.key]);
+    const usedKeys = allUsedKeyIdsInLayers(this.keyboard3?.layers);
+    const uniqueKeys = calculateUniqueKeys([...this.keyboard3.keys?.key]);
 
     for (let key of uniqueKeys) {
       if (!usedKeys.has(key.id)) {
@@ -159,25 +171,44 @@ export class KeysCompiler extends SectionCompiler {
         flags |= constants.keys_key_flags_notransform;
       }
       const id = sections.strs.allocString(key.id);
-      const longPress: ListItem = sections.list.allocListFromEscapedSpaces(
-        sections.strs,
-        // TODO-LDML: markers,variables
-        key.longPress
-      );
-      const longPressDefault = sections.strs.allocAndUnescapeString(
-        // TODO-LDML: variables
-        sections.vars.substituteMarkerString(key.longPressDefault),
-      );
-      const multiTap: ListItem = sections.list.allocListFromEscapedSpaces(
-        sections.strs,
-        // TODO-LDML: markers,variables
-        key.multiTap
-      );
+      const longPress: ListItem = sections.list.allocListFromSpaces(
+        key.longPress, {
+          stringVariables: true,
+          markers: true,
+          unescape: true,
+        },
+        sections);
+
+      const longPressDefault = sections.strs.allocString(key.longPressDefault,
+        {
+          stringVariables: true,
+          markers: true,
+          unescape: true,
+        },
+        sections);
+
+      const multiTap: ListItem = sections.list.allocListFromSpaces(
+        key.multiTap,
+        {
+          stringVariables: true,
+          markers: true,
+          unescape: true,
+        },
+        sections);
       const keySwitch = sections.strs.allocString(key.switch); // 'switch' is a reserved word
+
       const toRaw = key.to;
-      // TODO-LDML: variables
-      let toCooked = sections.vars.substituteMarkerString(toRaw);
-      const to = sections.strs.allocAndUnescapeString(toCooked, true);
+
+      let toCooked = sections.vars.substituteStrings(toRaw, sections);
+      toCooked = sections.vars.substituteMarkerString(toCooked);
+      const to = sections.strs.allocString(key.to,
+        {
+          stringVariables: true,
+          markers: true,
+          unescape: true,
+          singleOk: true
+        },
+        sections);
       if (!to.isOneChar) {
         flags |= constants.keys_key_flags_extend;
       }
@@ -194,6 +225,26 @@ export class KeysCompiler extends SectionCompiler {
         width,
       });
     }
+  }
+
+  private getKeymapFromForm(hardware : string, badScans?: Set<number>) : Constants.KeyMap {
+    return KeysCompiler.getKeymapFromForms(this.keyboard3?.forms.form, hardware, badScans);
+  }
+
+  public static getKeymapFromForms(forms: LDMLKeyboard.LKForm[], hardware: string, badScans?: Set<number>): Constants.KeyMap {
+    // seach in reverse form because of overrides
+    const ldmlForm = [...forms].reverse().find((f) => f.id === hardware);
+    if (!ldmlForm) {
+      return undefined;
+    }
+    return KeysCompiler.getKeymapFromScancodes(ldmlForm, badScans);
+  }
+
+  public static getKeymapFromScancodes(ldmlForm: LDMLKeyboard.LKForm, badScans?: Set<number>) {
+    const { scanCodes } = ldmlForm;
+    const ldmlScan = scanCodes.map(o => o.codes.split(" ").map(n => Number.parseInt(n, 16)));
+    const ldmlVkey = Constants.CLDRScanToKeyMap(ldmlScan, badScans);
+    return ldmlVkey;
   }
 
   /**
@@ -217,17 +268,24 @@ export class KeysCompiler extends SectionCompiler {
       valid = false;
     }
 
-    const keymap = Constants.HardwareToKeymap.get(hardware);
-    /* c8 ignore next 5 */
+    const badScans = new Set<number>();
+    const keymap = this.getKeymapFromForm(hardware, badScans);
     if (!keymap) {
-      // not reached due to XML validation
       this.callbacks.reportMessage(
         CompilerMessages.Error_InvalidHardware({ form: hardware })
       );
       valid = false;
+      return valid;
+    } else if (badScans.size !== 0) {
+      const codes = Array.from(badScans.values()).map(n => Number(n).toString(16)).sort();
+      this.callbacks.reportMessage(
+        CompilerMessages.Error_InvalidScanCode({ form: hardware, codes })
+      );
+      valid = false;
+      return valid;
     }
 
-    const uniqueKeys = calculateUniqueKeys([...this.keyboard.keys?.key]);
+    const uniqueKeys = calculateUniqueKeys([...this.keyboard3.keys?.key]);
     if (layer.row.length > keymap.length) {
       this.callbacks.reportMessage(
         CompilerMessages.Error_HardwareLayerHasTooManyRows()
@@ -287,7 +345,7 @@ export class KeysCompiler extends SectionCompiler {
     hardware: string
   ): Keys {
     const mod = translateLayerAttrToModifier(layer);
-    const keymap = Constants.HardwareToKeymap.get(hardware);
+    const keymap = this.getKeymapFromForm(hardware);
 
     let y = -1;
     for (let row of layer.row) {
@@ -300,7 +358,7 @@ export class KeysCompiler extends SectionCompiler {
 
         // TODO-LDML: we already validated that the key exists, above.
         // So here we only need the ID?
-        // let keydef = this.keyboard.keys?.key?.find(x => x.id == key);
+        // let keydef = this.keyboard3.keys?.key?.find(x => x.id == key);
 
         sect.kmap.push({
           vkey: keymap[y][x],